YinYang: Working implementation of three children.
With three elements, the trails are not overlapped
far enough for their shapes to be clean swirls.
There is still a small crevice on the inside of
the swirls. In order to cover the crevice, each
element is decorated with an additional
arc-outline in the front, adjusted by its angle
and with a very custom length.  The element behind
it joins this arc-outline with its own to complete
the outline, which is made just thick enough.

- Add decorator support to `Element`.

- Create the illusion of 'masking to reveal the
background` via backgroundColor.

- Minor renaming.